requirements.txt
pytest 
selenium 
allure-pytest 
requests